wp4

+++ tram-collision-avoidance/src/mainwindow.cpp
@@ -331,13 +331,13 @@ void MainWindow::processFrame(cv::Mat &iFrame)
         cv::rectangle(tVisualisation, mFeatures.tram, cv::Scalar(0, 255, 0), 1);
 
         //Draw pedestrians
-        for (int i = 0; i < mFeatures.pedestrians.size(); i++) {
+        for (size_t i = 0; i < mFeatures.pedestrians.size(); i++) {
             cv::Rect r = mFeatures.pedestrians[i];
             cv::rectangle(tVisualisation, r.tl(), r.br(), cv::Scalar(0,0,255), 2);
         }
 
         //Draw vehicles
-        for (int i = 0; i < mFeatures.vehicles.size(); i++) {
+        for (size_t i = 0; i < mFeatures.vehicles.size(); i++) {
             cv::Rect r = mFeatures.vehicles[i];
             cv::rectangle(tVisualisation, r.tl(), r.br(), cv::Scalar(255,0,0), 1);
         }
